Do not modify the request when X-Forwarded-Uri is not set
When the traefik-forward-auth server is behind an IngressRoute, the
X-Forwarded-Uri may not be set. If this is the case,
authentication.GetRequestURI will return a path of "/" which strips off
the "/_oauth" path before it can be handled on callback. If this header
is not set or is empty, then do not change the URL.